+++ doc/release-notes.md
@@ -1,1 +1,5 @@
 * Upgraded novaordis-utilities to a version higher then 4.2.2, which contained a OS implementation defect.
+
+* Initial support for a per-project configuration file (.nort/project.yaml)
+
+* Additional qualification sequence tests (installed version constraints).